# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /work-records:
    get:
      operationId: listWorkRecords
      tags: [employee-sections]
      summary: List work records across all employees in a single call
      description: |
        Aggregation endpoint that returns work records joined with employee
        data so the caller avoids fanning out one request per employee.
        All parameters are optional — omitting them returns all records.
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Inclusive lower bound on shiftDate (YYYY-MM-DD)
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Inclusive upper bound on shiftDate (YYYY-MM-DD)
        - name: shiftType
          in: query
          required: false
          schema:
            type: string
          description: Filter to a specific shift type value
        - name: departmentId
          in: query
          required: false
          schema:
            type: integer
          description: Filter to a specific department
        - name: employeeStatus
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/EmployeeStatus"
          description: Filter to employees with this status (omit for all statuses)
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Full-name or email substring search across employees
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 1-based page number
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: Number of records per page (max 200)
      responses:
        "200":
          description: Paginated work records with embedded employee context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkRecordsPage"
  /employees/{id}/work-records:
    get:
      operationId: listEmployeeWorkRecords
      tags: [employee-sections]
      summary: List work records (shifts/timesheets) for an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of work records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployeeWorkRecord"
    post:
      operationId: createEmployeeWorkRecord
      tags: [employee-sections]
      summary: Add a work record to an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeWorkRecordInput"
      responses:
        "201":
          description: Created work record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeWorkRecord"
  /employees/{id}/work-records/{recordId}:
    patch:
      operationId: updateEmployeeWorkRecord
      tags: [employee-sections]
      summary: Update a work record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: recordId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeWorkRecordUpdate"
      responses:
        "200":
          description: Updated work record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeWorkRecord"
        "404":
          description: Not found
    delete:
      operationId: deleteEmployeeWorkRecord
      tags: [employee-sections]
      summary: Delete a work record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: recordId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    EmployeeStatus:
      type: string
      enum: [active, inactive, on_leave, leaver]
...
    DepartmentSummary:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
...
    WorkRecordsPage:
      type: object
      required: [rows, total, page, pageSize, totalPages]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/WorkRecordRow"
        total:
          type: integer
          description: Total number of matching records across all pages
        page:
          type: integer
          description: Current 1-based page number
        pageSize:
          type: integer
          description: Number of records per page
        totalPages:
          type: integer
          description: Total number of pages (ceil(total / pageSize))
    WorkRecordRow:
      type: object
      required:
        - id
        - employeeId
        - shiftDate
        - shiftType
        - createdAt
        - employeeFirstName
        - employeeLastName
        - employeeEmail
        - employeeStatus
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        shiftDate:
          type: string
          format: date
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        hoursWorked:
          type: ["number", "null"]
        shiftType:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        employeeFirstName:
          type: string
        employeeLastName:
          type: string
        employeeEmail:
          type: string
        employeeStatus:
          $ref: "#/components/schemas/EmployeeStatus"
        employeeDepartments:
          type: array
          items:
            $ref: "#/components/schemas/DepartmentSummary"
        employeeAvatarUrl:
          type: ["string", "null"]
        employeeLeaverDate:
          type: ["string", "null"]
          format: date
          description: "Leaving date of the employee; null for active employees"
        employeeLeaverReason:
          type: ["string", "null"]
          description: "Leaver reason LOV value; null for active employees"
    EmployeeWorkRecord:
      type: object
      required: [id, employeeId, shiftDate, shiftType, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        shiftDate:
          type: string
          format: date
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        hoursWorked:
          type: ["number", "null"]
        shiftType:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    EmployeeWorkRecordInput:
      type: object
      required: [shiftDate]
      properties:
        shiftDate:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        hoursWorked:
          type: number
        shiftType:
          type: string
        notes:
          type: string
    EmployeeWorkRecordUpdate:
      type: object
      properties:
        shiftDate:
          type: string
          format: date
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        hoursWorked:
          type: ["number", "null"]
        shiftType:
          type: string
        notes:
          type: ["string", "null"]
...